Add github compile time environmental variables
When installing R dependencies as part of init.r, if they depend on
devtools::install_github and exist in a private repository, the
necessary and default GITHUB_PAT ENV variable is unavailable.

This commit looks for the $ENV_DIR directory for a file named
GITHUB_PAT. If it exists, it is made available just prior to execution
of init.r.